自定义404等待时间
