+++ pages/[prefix]/index.js
@@ -8,7 +8,7 @@ import { getNotion } from '@/lib/notion/getNotion'
 import { getPageTableOfContents } from '@/lib/notion/getPageTableOfContents'
 import { getLayoutByTheme } from '@/themes/theme'
 import md5 from 'js-md5'
-import { checkContainHttp, isBrowser } from '@/lib/utils'
+import { checkContainHttp } from '@/lib/utils'
 import { uploadDataToAlgolia } from '@/lib/algolia'
 import { siteConfig } from '@/lib/config'
 
@@ -20,7 +20,6 @@ import { siteConfig } from '@/lib/config'
  */
 const Slug = props => {
   const { post } = props
-  const router = useRouter()
 
   // 文章锁🔐
   const [lock, setLock] = useState(post?.password && post?.password !== '')
@@ -40,20 +39,6 @@ const Slug = props => {
 
   // 文章加载
   useEffect(() => {
-    // 404
-    if (!post) {
-      setTimeout(() => {
-        if (isBrowser) {
-          const article = document.getElementById('notion-article')
-          if (!article) {
-            router.push('/404').then(() => {
-              console.warn('找不到页面', router.asPath)
-            })
-          }
-        }
-      }, 8 * 1000) // 404时长 8秒
-    }
-
     // 文章加密
     if (post?.password && post?.password !== '') {
       setLock(true)
