+++ themes/medium/index.js
@@ -156,6 +156,23 @@ const LayoutSlug = props => {
         </div>
   )
 
+  const router = useRouter()
+  useEffect(() => {
+    // 404
+    if (!post) {
+      setTimeout(() => {
+        if (isBrowser) {
+          const article = document.getElementById('notion-article')
+          if (!article) {
+            router.push('/404').then(() => {
+              console.warn('找不到页面', router.asPath)
+            })
+          }
+        }
+      }, siteConfig('POST_WAITING_TIME_FOR_404') * 1000)
+    }
+  }, [post])
+
   return (
         <div showInfoCard={true} slotRight={slotRight} {...props} >
             {/* 文章锁 */}
