+++ themes/simple/index.js
@@ -12,6 +12,7 @@ import dynamic from 'next/dynamic'
 import NotionPage from '@/components/NotionPage'
 import AlgoliaSearchModal from '@/components/AlgoliaSearchModal'
 import { LAYOUT_MAPPINGS } from '@/blog.config'
+import { useRouter } from 'next/router'
 
 // 主题组件
 const BlogListScroll = dynamic(() => import('./components/BlogListScroll'), { ssr: false });
@@ -217,6 +218,23 @@ const LayoutSlug = props => {
  * @returns
  */
 const Layout404 = (props) => {
+  const { post } = props
+  const router = useRouter()
+  useEffect(() => {
+    // 404
+    if (!post) {
+      setTimeout(() => {
+        if (isBrowser) {
+          const article = document.getElementById('notion-article')
+          if (!article) {
+            router.push('/404').then(() => {
+              console.warn('找不到页面', router.asPath)
+            })
+          }
+        }
+      }, siteConfig('POST_WAITING_TIME_FOR_404') * 1000)
+    }
+  }, [post])
   return <>
         404 Not found.
     </>
